requirements.txt
yfinance
pandas
scikit-learn
requests
numpy
python-dotenv
streamlit
plotly
